fix: type unwrapNullableUnionType for ASTDescribedColumnType
Move the helper into ast-describe where the union type is defined to
fix generate package typecheck (TS2345).
